extract StacHrefBuilder class into lib/stac.py
