+++ build_stac_items.py
@@ -182,7 +182,7 @@ def build_strip_stac_item(base_url, domain, raster):
     mask_info = stac.RasterAssetInfo.from_raster(raster.bitmask)
     matchtag_info = stac.RasterAssetInfo.from_raster(raster.matchtag)
 
-    href_builder = StacHrefBuilder(
+    href_builder = stac.StacHrefBuilder(
         base_url=base_url, s3_bucket=S3_BUCKET, domain=domain, raster=raster
     )
 
@@ -396,7 +396,7 @@ def build_mosaic_stac_item(base_url, domain, tile):
     maxdate_info = stac.RasterAssetInfo.from_raster(tile.maxdate)
     mindate_info = stac.RasterAssetInfo.from_raster(tile.mindate)
 
-    href_builder = StacHrefBuilder(
+    href_builder = stac.StacHrefBuilder(
         base_url=base_url, s3_bucket=S3_BUCKET, domain=domain, raster=tile
     )
 
@@ -613,7 +613,7 @@ def build_mosaic_v3_stac_item(base_url, domain, tile):
     # Get info for each raster asset, except "browse" which is done conditionally later
     dem_info = stac.RasterAssetInfo.from_raster(tile.srcfp)
 
-    href_builder = StacHrefBuilder(
+    href_builder = stac.StacHrefBuilder(
         base_url=base_url, s3_bucket=S3_BUCKET, domain=domain, raster=tile
     )
 
@@ -768,58 +768,6 @@ def iso8601(date_time, msg=""):
 
     logger.error(f"null date: {msg}")
     return None
-
-
-class StacHrefBuilder:
-    def __init__(
-            self, base_url: str, s3_bucket: str, domain: str, raster: dem.SetsmDem | dem.SetsmTile
-    ):
-        self.base_url = base_url.rstrip("/") # Remove trailing slash if present
-        self.base_s3_url = f"s3://{s3_bucket}"
-
-        if isinstance(raster, dem.SetsmDem):
-            kind = "strips"
-            geocell_or_supertile = raster.geocell
-            release_version = raster.release_version
-            res_str = raster.res_str
-            item_id = raster.stripid
-        elif isinstance(raster, dem.SetsmTile):
-            kind = "mosaics"
-            geocell_or_supertile = raster.supertile_id_no_res
-            release_version = f"v{raster.release_version}"
-            res_str = raster.res
-            item_id = raster.tileid
-        else:
-            raise ValueError(
-                f"raster argument must be either {type(dem.SetsmDem)} or {type(dem.SetsmTile)}. Got {type(raster)}"
-            )
-
-        self._partial_asset_key = f"{domain}/{kind}/{release_version}/{res_str}/{geocell_or_supertile}"
-        self._item_key = f"{domain}/{kind}/{release_version}/{res_str}/{geocell_or_supertile}/{item_id}.json"
-        self._catalog_key = f"{domain}/{kind}/{release_version}/{res_str}/{geocell_or_supertile}.json"
-        self._collection_key = f"{domain}/{kind}/{release_version}/{res_str}.json"
-        self._root_key = "pgc-data-stac.json"
-
-    def item_href(self, as_s3: bool = False) -> str:
-        base = self.base_s3_url if as_s3 else self.base_url
-        return f"{base}/{self._item_key}"
-
-    def catalog_href(self, as_s3: bool = False) -> str:
-        base = self.base_s3_url if as_s3 else self.base_url
-        return f"{base}/{self._catalog_key}"
-
-    def collection_href(self, as_s3: bool = False) -> str:
-        base = self.base_s3_url if as_s3 else self.base_url
-        return f"{base}/{self._collection_key}"
-
-    def root_href(self, as_s3: bool = False) -> str:
-        base = self.base_s3_url if as_s3 else self.base_url
-        return f"{base}/{self._root_key}"
-
-    def asset_href(self, filepath: str | pathlib.Path, as_s3: bool = False) -> str:
-        filename = pathlib.Path(filepath).name if isinstance(filepath, str) else filepath.name
-        base = self.base_s3_url if as_s3 else self.base_url
-        return f"{base}/{self._partial_asset_key}/{filename}"
 
 
 if __name__ == '__main__':
